Mejoras en el codigo, correcion metodo error y se agrega sql para cargar datos por defecto en los test

+++ Technical-Challenges/Rest-Library-Services/src/main/java/com/hackerrank/biblioteca/controller/PrestamoControlador.java
@@ -30,6 +30,7 @@ import com.hackerrank.biblioteca.service.PrestamoService;
 import com.hackerrank.biblioteca.service.UsuarioService;
 
 
+
 @RestController
 @RequestMapping("prestamo")
 public class PrestamoControlador {
@@ -141,7 +142,7 @@ public class PrestamoControlador {
 
 
 	@GetMapping("/{id}")
-	public ResponseEntity<PrestamoResponseDTOByID> getPrestamosxID (@PathVariable("id") Long id) {
+	public ResponseEntity<PrestamoResponseDTOByID> getPrestamosxID (@PathVariable("id") int id) {
 
 		ModelMapper modelMapper = new ModelMapper();
 		modelMapper.getConfiguration().setAmbiguityIgnored(true);
