+++ Technical-Challenges/Rest-Library-Services/src/main/java/com/hackerrank/biblioteca/repository/ILibroRepository.java
@@ -6,8 +6,11 @@ import org.springframework.stereotype.Repository;
 import com.hackerrank.biblioteca.model.Libro;
 
 
+
+
+
 @Repository
-public interface ILibroRepository extends CrudRepository<Libro, Long>{
+public interface ILibroRepository extends CrudRepository<Libro, Integer>{
 	
 
 }
